oteldemo: lint fixes across the new APM pages
Drive-by cleanup after running eslint on the four new commits.

- Deterministic skeleton widths: Math.random() is flagged by
  react-hooks/purity. Swap random widths for hard-coded arrays —
  still varied, still shimmery, still lint-clean.
- HomePage: Date.now() moved into a useState lazy initializer so it
  runs once at mount, not on every render.
- TimeRangePicker: constants moved to a new timeRanges.ts module so
  the component file only exports the component (satisfies
  react-refresh/only-export-components).

No behavior changes.
